requirements.txt
httpx ~= 0.28.1
chess ~= 1.11.2
PyYAML ~= 6.0
